(#143) 기존 메인 화면 피드 화면으로 교체
