+++ app/src/main/java/com/boostcamp/travery/course/list/adapter/CourseListAdapter.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.travery.main.adapter
+package com.boostcamp.travery.course.list.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,11 +7,11 @@ import androidx.databinding.ObservableArrayList
 import com.boostcamp.travery.R
 import com.boostcamp.travery.base.BaseViewHolder
 import com.boostcamp.travery.base.ObservableRecyclerViewAdapter
+import com.boostcamp.travery.course.list.adapter.viewholder.CourseViewHolder
+import com.boostcamp.travery.course.list.adapter.viewholder.GroupItem
+import com.boostcamp.travery.course.list.adapter.viewholder.GroupViewHolder
 import com.boostcamp.travery.databinding.ItemCourseListBinding
 import com.boostcamp.travery.databinding.ItemCourseListGroupBinding
-import com.boostcamp.travery.main.adapter.viewholder.CourseViewHolder
-import com.boostcamp.travery.main.adapter.viewholder.GroupItem
-import com.boostcamp.travery.main.adapter.viewholder.GroupViewHolder
 
 class CourseListAdapter(private val data: ObservableArrayList<Any>) : ObservableRecyclerViewAdapter<Any, BaseViewHolder>(data) {
