+++ app/src/main/res/layout/main_feed.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:bind="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:bind="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools">
 
     <data>
 
         <variable
             name="viewmodel"
-            type="com.boostcamp.travery.main.MainViewModel" />
+            type="com.boostcamp.travery.feed.NewsFeedViewModel"/>
     </data>
 
     <androidx.drawerlayout.widget.DrawerLayout
@@ -23,22 +23,22 @@
             android:layout_height="match_parent">
 
             <com.google.android.material.appbar.AppBarLayout
+                style="@style/AppTheme.AppBarOverlay"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                style="@style/AppTheme.AppBarOverlay">
+                android:layout_height="wrap_content">
 
                 <include
                     android:id="@+id/toolBar"
-                    layout="@layout/toolbar" />
+                    layout="@layout/toolbar"/>
             </com.google.android.material.appbar.AppBarLayout>
 
             <include
-                android:id="@+id/app_bar_main"
-                layout="@layout/app_bar_main"
+                android:id="@+id/main_layout_newsfeed"
+                layout="@layout/main_layout_newsfeed"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 app:layout_behavior="@string/appbar_scrolling_view_behavior"
-                bind:viewmodel="@{viewmodel}" />
+                bind:viewModel="@{viewmodel}"/>
         </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
         <com.google.android.material.navigation.NavigationView
@@ -47,6 +47,6 @@
             android:layout_height="match_parent"
             android:layout_gravity="start"
             app:headerLayout="@layout/nav_header_main"
-            app:menu="@menu/activity_main_drawer" />
+            app:menu="@menu/activity_main_drawer"/>
     </androidx.drawerlayout.widget.DrawerLayout>
 </layout>
